move updateRoles method to user-authorization service
