+++ src/app/core/user/profile/user-profile.controller.js
@@ -15,6 +15,7 @@ const
 	User = dbs.admin.model('User'),
 
 	resourcesService = require('../../resources/resources.service')(),
+	userAuthorizationService = require('../auth/user-authorization.service'),
 	userProfileService = require('./user-profile.service'),
 	newUserEmailService = require('../new-user-email.service');
 
@@ -78,29 +79,6 @@ function searchUsers(req, res, copyUserFn) {
 	});
 }
 
-exports.updateRoles = (user, authConfig) => {
-	const strategy = _.get(authConfig, 'roleStrategy', 'local');
-	const isHybrid = strategy === 'hybrid';
-
-	if (isHybrid) {
-		user.localRoles = user.roles;
-	}
-	if (strategy === 'external' || isHybrid) {
-		let updatedRoles = {};
-		let externalRoles = user.externalRoles || [];
-		let externalRoleMap = authConfig.externalRoleMap;
-
-		let keys = _.keys(externalRoleMap);
-
-		keys.forEach((key) => {
-			updatedRoles[key] = (isHybrid && user.roles && user.roles[key]) || externalRoles.indexOf(externalRoleMap[key]) !== -1;
-		});
-
-		user.roles = updatedRoles;
-	}
-};
-
-
 /**
  * Standard User Operations
  */
@@ -120,7 +98,7 @@ exports.getCurrentUser = (req, res) => {
 
 	let userCopy = User.fullCopy(user);
 
-	exports.updateRoles(userCopy, config.auth);
+	userAuthorizationService.updateRoles(userCopy, config.auth);
 
 	res.status(200).json(userCopy);
 };
