+++ src/app/core/user/profile/user-profile.controller.spec.js
@@ -175,90 +175,4 @@ describe('User Profile Controller:', () => {
 
 	});
 
-	describe('updateRoles', () => {
-		it('roleStrategy === local; should pass through roles as is', () => {
-			let user = {
-				roles: {
-					user: true,
-					contentSteward: true,
-					equipmentSteward: false
-				}
-			};
-
-			userProfileController.updateRoles(user, {
-				roleStrategy: 'local'
-			});
-
-			user.should.be.an.Object();
-			user.roles.should.be.an.Object();
-			user.roles.user.should.be.true();
-			user.roles.contentSteward.should.be.true();
-			user.roles.equipmentSteward.should.be.false();
-
-			should.not.exist(user.localRoles);
-		});
-
-		it('roleStrategy === external; should pass through roles as is', () => {
-			let user = {
-				roles: {
-					user: true,
-					contentSteward: true,
-					equipmentSteward: false
-				},
-				externalRoles: ['USER', 'EQUIPMENT_STEWARD']
-			};
-
-			userProfileController.updateRoles(user, {
-				roleStrategy: 'external',
-				externalRoleMap: {
-					user: 'USER',
-					contentSteward: 'CONTENT_STEWARD',
-					equipmentSteward: 'EQUIPMENT_STEWARD',
-					admin: 'ADMIN'
-				}
-			});
-
-			user.should.be.an.Object();
-			user.roles.should.be.an.Object();
-			user.roles.user.should.be.true();
-			user.roles.contentSteward.should.be.false();
-			user.roles.equipmentSteward.should.be.true();
-
-			should.not.exist(user.localRoles);
-		});
-
-		it('roleStrategy === hybrid; should pass through roles as is', () => {
-			let user = {
-				roles: {
-					user: true,
-					contentSteward: true,
-					equipmentSteward: false
-				},
-				externalRoles: ['USER', 'EQUIPMENT_STEWARD']
-			};
-
-			userProfileController.updateRoles(user, {
-				roleStrategy: 'hybrid',
-				externalRoleMap: {
-					user: 'USER',
-					contentSteward: 'CONTENT_STEWARD',
-					equipmentSteward: 'EQUIPMENT_STEWARD',
-					admin: 'ADMIN'
-				}
-			});
-
-			user.should.be.an.Object();
-			user.roles.should.be.an.Object();
-			user.roles.user.should.be.true();
-			user.roles.contentSteward.should.be.true();
-			user.roles.equipmentSteward.should.be.true();
-			user.roles.admin.should.be.false();
-
-			user.localRoles.should.be.an.Object();
-			user.localRoles.user.should.be.true();
-			user.localRoles.contentSteward.should.be.true();
-			user.localRoles.equipmentSteward.should.be.false();
-		});
-	});
-
 });
